Refactor Runner out of __init__
Reduces amount of code we have in package __init__.

Partial: #812
