+++ test/TestImportIncludeRole.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ansiblelint import Runner
+from ansiblelint.runner import Runner
 
 
 ROLE_TASKS_MAIN = '''
