+++ test/TestWithSkipTagId.py
@@ -1,5 +1,6 @@
 import unittest
-from ansiblelint import RulesCollection, Runner
+from ansiblelint import RulesCollection
+from ansiblelint.runner import Runner
 from ansiblelint.rules.TrailingWhitespaceRule import TrailingWhitespaceRule
